+++ test/include.test.js
@@ -36,35 +36,6 @@ describe(Support.getTestDialectTeaser("Include"), function () {
       })
     })
 
-    // We don't support naming associations the same as the foreign key, however the system should not crash because of it, the results hould just be wrong as is expected behaviour currently.
-    it('should not throw an error when an empty include is named the same as the foreign key', function (done) {
-      var section = this.sequelize.define('section', { name: DataTypes.STRING });
-      var layout = this.sequelize.define('layout', { name: DataTypes.STRING });
-
-      section.belongsTo(layout, {
-        as: layout.name,
-        foreignKey: layout.name,
-        foreignKeyConstraint: true
-      });
-       
-      this.sequelize.sync({force: true}).done(function(err) {
-        expect(err).to.be.null
-
-        section.create({ name: 'test1' }).success(function() {
-          section.find({
-            where: { name: 'test1' },
-            include: [
-              { model: layout, as: 'layout'}
-            ]
-          }).done(function(err, user) {
-            expect(err).to.be.null
-            expect(user).to.be.ok
-            done()
-          });
-        });
-      });
-    })
-
     it('should support a empty hasOne include', function (done) {
       var Company = this.sequelize.define('Company', {})
         , Person = this.sequelize.define('Person', {})